# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CryptoX Exchange API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change current user's password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "200":
          description: Password changed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized or wrong current password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/kyc/verify:
    post:
      operationId: submitKyc
      tags: [auth]
      summary: Submit KYC verification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KycSubmitRequest"
      responses:
        "200":
          description: KYC submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        name:
          type: string
        phone:
          type: string
        country:
          type: string
        dateOfBirth:
          type: string
        timezone:
          type: string
          description: Current IANA timezone reported by the user's device, such as Africa/Lagos.
        referralCode:
          type: string
          minLength: 3
          maxLength: 64
      required:
        - email
        - password
        - name
        - phone
        - country
        - dateOfBirth
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        timezone:
          type: string
          description: Current IANA timezone reported by the user's device, such as Africa/Lagos.
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        message:
          type: string
      required:
        - user
        - message
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        experience:
          type: string
        usdBalance:
          type: number
        kycStatus:
          type: string
          enum: [unverified, pending, verified, rejected]
        role:
          type: string
          enum: [user, admin]
        status:
          type: string
          enum: [active, suspended]
        createdAt:
          type: string
        referralCode:
          type: string
      required:
        - id
        - email
        - name
        - experience
        - usdBalance
        - kycStatus
        - role
        - status
        - createdAt
        - referralCode
...
    ChangePasswordRequest:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
      required:
        - currentPassword
        - newPassword
    KycSubmitRequest:
      type: object
      properties:
        fullName:
          type: string
          minLength: 2
          maxLength: 200
        dateOfBirth:
          type: string
          pattern: '^\d{4}-\d{2}-\d{2}$'
        country:
          type: string
          minLength: 2
          maxLength: 100
        idNumber:
          type: string
          minLength: 2
          maxLength: 200
      required:
        - fullName
        - dateOfBirth
        - country
        - idNumber
...