# lib/api-zod/src/generated/types/registerRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CryptoX Exchange API
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterRequest {
  email: string;
  password: string;
  name: string;
  phone: string;
  country: string;
  dateOfBirth: string;
  /** Current IANA timezone reported by the user's device, such as Africa/Lagos. */
  timezone?: string;
  /**
   * @minLength 3
   * @maxLength 64
   */
  referralCode?: string;
}