# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CryptoX Exchange API
 * OpenAPI spec version: 0.1.0
 */
import type { UserKycStatus } from "./userKycStatus";
import type { UserRole } from "./userRole";
import type { UserStatus } from "./userStatus";

export interface User {
  id: number;
  email: string;
  name: string;
  experience: string;
  usdBalance: number;
  kycStatus: UserKycStatus;
  role: UserRole;
  status: UserStatus;
  createdAt: string;
  referralCode: string;
}